--- messaging/value/.gitignore
@@ -1,12 +1,0 @@
-node_modules
-.env
-coverage
-coverage.json
-typechain
-typechain-types
-
-# Hardhat files
-cache
-artifacts
-
-access_token
