--- messaging/value/contracts/test/TestContracts.sol
@@ -1,5 +1,0 @@
-// SPDX-License-Identifier: MIT
-pragma solidity 0.8.7;
-
-import "@zetachain/toolkit/contracts/ZetaConnectorMock.sol";
-import "@zetachain/toolkit/contracts/EthZetaMock.sol";
